Developer

+++ src/main/java/pe/edu/upc/micasita/Micasitaplataform/Property/domain/model/commands/ChangePropertyDescriptionCommand.java
@@ -1,4 +1,4 @@
 package pe.edu.upc.micasita.Micasitaplataform.Property.domain.model.commands;
 
-public record ChangePropertyDescriptionCommand(Integer id, String newDescription) {
+public record ChangePropertyDescriptionCommand(Long id, String newDescription) {
 }
