+++ src/main/java/pe/edu/upc/micasita/Micasitaplataform/Property/domain/model/commands/ChangePropertyPriceCommand.java
@@ -1,4 +1,4 @@
 package pe.edu.upc.micasita.Micasitaplataform.Property.domain.model.commands;
 
-public record ChangePropertyPriceCommand(Integer id, Long newPrice) {
+public record ChangePropertyPriceCommand(Long id, Long newPrice) {
 }
