+++ src/main/java/pe/edu/upc/micasita/Micasitaplataform/Property/domain/model/commands/ChangePropertyStatusCommand.java
@@ -1,5 +1,5 @@
 package pe.edu.upc.micasita.Micasitaplataform.Property.domain.model.commands;
 
 
-public record ChangePropertyStatusCommand(Integer id, String status) {
+public record ChangePropertyStatusCommand(Long id, String status) {
 }
