+++ src/main/java/pe/edu/upc/micasita/Micasitaplataform/Property/domain/model/commands/ChangePropertyTypeCommand.java
@@ -1,4 +1,4 @@
 package pe.edu.upc.micasita.Micasitaplataform.Property.domain.model.commands;
 
-public record ChangePropertyTypeCommand(Integer id, String newType) {
+public record ChangePropertyTypeCommand(Long id, String newType) {
 }
